+++ pom.xml
@@ -6,7 +6,7 @@
     <groupId>com.howarth.cloud</groupId>
 
     <artifactId>home</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
+    <version>LOCAL-SNAPSHOT</version>
     <packaging>war</packaging>
 
     <parent>
