+++ src/main/java/com/howarth/cloudplatform/config/StorageProperties.java
@@ -1,0 +1,17 @@
+package com.howarth.cloudplatform.config;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties("storage")
+public class StorageProperties {
+    private String location = "./ROOT/upload-dir/";
+
+    public String getLocation() {
+        return location;
+    }
+
+    public void setLocation(String location) {
+        this.location = location;
+    }
+
+}
