+++ src/main/java/com/howarth/cloudplatform/exception/StorageException.java
@@ -1,0 +1,12 @@
+package com.howarth.cloudplatform.exception;
+
+public class StorageException extends RuntimeException {
+
+    public StorageException(String message) {
+        super(message);
+    }
+
+    public StorageException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
