+++ src/main/java/com/howarth/cloudplatform/exception/StorageFileNotFoundException.java
@@ -1,0 +1,12 @@
+package com.howarth.cloudplatform.exception;
+
+public class StorageFileNotFoundException extends StorageException {
+
+    public StorageFileNotFoundException(String message) {
+        super(message);
+    }
+
+    public StorageFileNotFoundException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
