--- src/main/java/com/howarth/cloudplatform/fileuploads/storage/StorageProperties.java
@@ -1,17 +1,0 @@
-package com.howarth.cloudplatform.fileuploads.storage;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties("storage")
-public class StorageProperties {
-    private String location = "./ROOT/upload-dir/";
-
-    public String getLocation() {
-        return location;
-    }
-
-    public void setLocation(String location) {
-        this.location = location;
-    }
-
-}
