--- src/main/java/com/howarth/cloudplatform/fileuploads/storage/StorageService.java
@@ -1,14 +1,0 @@
-package com.howarth.cloudplatform.fileuploads.storage;
-
-import org.springframework.web.multipart.MultipartFile;
-
-import java.nio.file.Path;
-
-public interface StorageService {
-
-    void init();
-
-    void store(MultipartFile file);
-
-    Path load(String filename);
-}
