--- src/main/java/com/howarth/cloudplatform/fileuploads/storage/database/ApplicationApp.java
@@ -1,56 +1,0 @@
-package com.howarth.cloudplatform.fileuploads.storage.database;
-
-import javax.persistence.*;
-
-@Entity
-public class ApplicationApp {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    @Column(unique = true)
-    private String name;
-
-    private String description;
-
-    private String logo;
-
-    private String username;
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getLogo() {
-        return logo;
-    }
-
-    public void setLogo(String logo) {
-        this.logo = logo;
-    }
-}
