--- src/main/java/com/howarth/cloudplatform/fileuploads/storage/exception/StorageException.java
@@ -1,12 +1,0 @@
-package com.howarth.cloudplatform.fileuploads.storage.exception;
-
-public class StorageException extends RuntimeException {
-
-    public StorageException(String message) {
-        super(message);
-    }
-
-    public StorageException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
