--- src/main/java/com/howarth/cloudplatform/fileuploads/storage/exception/StorageFileNotFoundException.java
@@ -1,12 +1,0 @@
-package com.howarth.cloudplatform.fileuploads.storage.exception;
-
-public class StorageFileNotFoundException extends StorageException {
-
-    public StorageFileNotFoundException(String message) {
-        super(message);
-    }
-
-    public StorageFileNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
