--- src/main/java/com/howarth/cloudplatform/library/database/DiamondRepository.java
@@ -1,10 +1,0 @@
-package com.howarth.cloudplatform.library.database;
-
-import com.howarth.cloudplatform.library.database.model.DiamondStatistic;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface DiamondRepository extends JpaRepository<DiamondStatistic, Long> {
-    DiamondStatistic findTopByOrderByIdDesc();
-}
