+++ src/main/java/com/howarth/cloudplatform/model/ValidUseToken.java
@@ -1,0 +1,28 @@
+package com.howarth.cloudplatform.model;
+
+public class ValidUseToken {
+
+    private String username;
+    private boolean valid;
+
+    public ValidUseToken(String username, boolean valid) {
+        this.username = username;
+        this.valid = valid;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public boolean isValid() {
+        return valid;
+    }
+
+    public void setValid(boolean valid) {
+        this.valid = valid;
+    }
+}
