+++ src/main/java/com/howarth/cloudplatform/model/VerifiedToken.java
@@ -1,0 +1,31 @@
+package com.howarth.cloudplatform.model;
+
+public class VerifiedToken {
+
+    private String username;
+    private boolean verified;
+
+    public VerifiedToken(String username, boolean verified) {
+        this.username = username;
+        this.verified = verified;
+    }
+
+    public VerifiedToken() {
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public boolean isVerified() {
+        return verified;
+    }
+
+    public void setVerified(boolean verified) {
+        this.verified = verified;
+    }
+}
