+++ src/main/java/com/howarth/cloudplatform/service/StorageService.java
@@ -1,0 +1,14 @@
+package com.howarth.cloudplatform.service;
+
+import org.springframework.web.multipart.MultipartFile;
+
+import java.nio.file.Path;
+
+public interface StorageService {
+
+    void init();
+
+    void store(MultipartFile file);
+
+    Path load(String filename);
+}
