+++ src/main/java/com/howarth/cloudplatform/service/UserDetailsServiceImpl.java
@@ -1,0 +1,29 @@
+package com.howarth.cloudplatform.service;
+
+import com.howarth.cloudplatform.model.ApplicationUser;
+import com.howarth.cloudplatform.dao.ApplicationUserDao;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+import static java.util.Collections.emptyList;
+
+@Service
+public class UserDetailsServiceImpl implements UserDetailsService {
+    private ApplicationUserDao applicationUserDao;
+
+    public UserDetailsServiceImpl(ApplicationUserDao applicationUserDao) {
+        this.applicationUserDao = applicationUserDao;
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        ApplicationUser applicationUser = applicationUserDao.findByUsername(username);
+        if (applicationUser == null) {
+            throw new UsernameNotFoundException(username);
+        }
+        return new User(applicationUser.getUsername(), applicationUser.getPassword(), emptyList());
+    }
+}
