--- src/main/java/com/howarth/cloudplatform/user/database/ApplicationUser.java
@@ -1,36 +1,0 @@
-package com.howarth.cloudplatform.user.database;
-
-import javax.persistence.*;
-
-@Entity
-public class ApplicationUser {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    @Column(unique = true)
-    private String username;
-
-    private String password;
-
-    public long getId() {
-        return id;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
