--- src/main/java/com/howarth/cloudplatform/user/database/ApplicationUserRepository.java
@@ -1,7 +1,0 @@
-package com.howarth.cloudplatform.user.database;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ApplicationUserRepository extends JpaRepository<ApplicationUser, Long> {
-    ApplicationUser findByUsername(String username);
-}
