--- src/main/java/com/howarth/cloudplatform/virtualbank/ValidUseToken.java
@@ -1,28 +1,0 @@
-package com.howarth.cloudplatform.virtualbank;
-
-public class ValidUseToken {
-
-    private String username;
-    private boolean valid;
-
-    public ValidUseToken(String username, boolean valid) {
-        this.username = username;
-        this.valid = valid;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public boolean isValid() {
-        return valid;
-    }
-
-    public void setValid(boolean valid) {
-        this.valid = valid;
-    }
-}
