--- src/main/java/com/howarth/cloudplatform/virtualbank/database/BankAccountRepository.java
@@ -1,10 +1,0 @@
-package com.howarth.cloudplatform.virtualbank.database;
-
-import com.howarth.cloudplatform.virtualbank.database.model.BankAccount;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface BankAccountRepository extends JpaRepository<BankAccount, Long> {
-    BankAccount findByUsername(String username);
-}
