--- src/main/java/com/howarth/cloudplatform/virtualbank/database/model/BankAccount.java
@@ -1,45 +1,0 @@
-package com.howarth.cloudplatform.virtualbank.database.model;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-public class BankAccount {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private long id;
-
-    private String username;
-    private int balance;
-
-    public BankAccount() {
-    }
-
-    public BankAccount(String username, int balance) {
-        this.username = username;
-        this.balance = balance;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public int getBalance() {
-        return balance;
-    }
-
-    public void setBalance(int balance) {
-        this.balance = balance;
-    }
-
-    public long getId() {
-        return id;
-    }
-}
